utils: improve error handing in ensureDirectorySync
If after `mkdirSync()` the directory exists, I think it's the definition of
success, and we shouldn't care about errors in that case.

We could also go the other way, check if the folder exists before creating it.
I guess which version makes fewer syscalls would depend on usage.

I also think it's important to throw the original error if we can't fix it.
For example, if we got a permission error, I don't want to attempt creating the
parent directory, and then to return the error about the parent if that fails
as well.

Because of `resolve(directoryPath)` at the API boundary it's not possible to
match Node's behavior for relative paths. Maybe we should resolve only when
we need to go outside of cwd.