breaking!: deprecate/remove @file-services/webpack we're not using it and it's a blocker for memory-fs to support binary read/write.
We could write a Buffer.from() specifically in the test, but that would mean that writing strings to the overlayed memory-fs will make webpack throw. those get converted to Uint8Array, while webpack expects and verifies it either a string or a fully-pledged Buffer.

this previously worked because memory-fs forced strings to be returned, even when not given an encoding.